reorg xpath3 coverage tests into themed owners
