+++ xpath3/format_number_test.go
