+++ xpath3/functions_json_test.go
