+++ xpath3/functions_serialize_test.go
