+++ xpath3/functions_test.go
@@ -1530,3 +1530,26 @@ func TestFnIndexOfSearchCardinality(t *testing.T) {
 		require.Equal(t, 0, seq.Len())
 	})
 }
+
+func TestBuiltinFunctionQueries(t *testing.T) {
+	require.True(t, xpath3.IsBuiltinFunction("abs"))
+	require.False(t, xpath3.IsBuiltinFunction("definitely-not-a-builtin"))
+
+	require.True(t, xpath3.IsBuiltinFunctionNS(xpath3.NSFn, "count"))
+	require.False(t, xpath3.IsBuiltinFunctionNS("urn:nope", "count"))
+
+	require.True(t, xpath3.BuiltinFunctionAcceptsArity(xpath3.NSFn, "abs", 1))
+	require.False(t, xpath3.BuiltinFunctionAcceptsArity(xpath3.NSFn, "abs", 5))
+	require.False(t, xpath3.BuiltinFunctionAcceptsArity(xpath3.NSFn, "nope", 1))
+}
+
+func TestPredeclaredNamespaces(t *testing.T) {
+	ns := xpath3.PredeclaredNamespaces()
+	require.Equal(t, xpath3.NSFn, ns["fn"])
+	require.NotEmpty(t, ns["xs"])
+
+	// Mutating the returned copy must not affect package state.
+	ns["fn"] = "tampered"
+	ns2 := xpath3.PredeclaredNamespaces()
+	require.Equal(t, xpath3.NSFn, ns2["fn"])
+}
