+++ xpath3/static_check_test.go
